Moved git stuff again, changed to core to actual core functions, and owner to just owner functions
